requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
aioredis==2.0.1
    # via flask-mailing
aiosmtplib==1.1.7
    # via flask-mailing
anyio==3.6.1
    # via httpcore
asgiref==3.5.2
    # via flask-mailing
aspose-words==22.9.0
    # via -r requirements.in
async-timeout==4.0.2
    # via aioredis
babel==2.10.3
    # via flask-babel
beautifulsoup4==4.11.1
    # via -r requirements.in
blinker==1.5
    # via flask-mailing
cachelib==0.9.0
    # via flask-session2
certifi==2022.9.24
    # via
    #   httpcore
    #   httpx
click==8.1.3
    # via flask
dnspython==2.2.1
    # via email-validator
email-validator==1.3.0
    # via flask-mailing
flask==2.2.2
    # via
    #   -r requirements.in
    #   flask-babel
    #   flask-mailing
    #   flask-session2
    #   flask-sqlalchemy
flask-babel==2.0.0
    # via -r requirements.in
flask-mailing==0.2.0
    # via -r requirements.in
flask-session2==1.3.1
    # via -r requirements.in
flask-sqlalchemy==3.0.2
    # via -r requirements.in
greenlet==1.1.3.post0
    # via sqlalchemy
h11==0.12.0
    # via httpcore
httpcore==0.15.0
    # via httpx
httpx==0.23.0
    # via flask-mailing
idna==3.4
    # via
    #   anyio
    #   email-validator
    #   rfc3986
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via
    #   flask
    #   flask-babel
libsass==0.21.0
    # via -r requirements.in
markupsafe==2.1.1
    # via
    #   jinja2
    #   werkzeug
pydantic==1.10.2
    # via flask-mailing
python-dotenv==0.21.0
    # via -r requirements.in
pytz==2022.4
    # via
    #   babel
    #   flask-babel
    #   flask-session2
rfc3986[idna2008]==1.5.0
    # via httpx
six==1.16.0
    # via libsass
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
soupsieve==2.3.2.post1
    # via beautifulsoup4
sqlalchemy==1.4.42
    # via
    #   -r requirements.in
    #   flask-sqlalchemy
typing-extensions==4.4.0
    # via
    #   aioredis
    #   flask-mailing
    #   pydantic
werkzeug==2.2.2
    # via flask
